requirements.txt
discord.py
os
random
